requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dataclasses==0.8
Flask==2.0.1
future==0.18.2
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
